refactor[#122] : dogBreeds 파일 위치 변경
