--- src/features/auth/constants/information.ts
@@ -1,159 +1,0 @@
-export const dogBreeds = [
-  "그레이 하운드",
-  "그레이트 데인",
-  "그레이트 피레네",
-  "골든 리트리버",
-  "고든 세터",
-  "글렌 오브 이말 테리어",
-  "노르웨이 엘크하운드",
-  "노르웨이 버훈트",
-  "노바 스코셔 덕 톨링 레트리버",
-  "노퍽 테리어",
-  "노리치 테리어",
-  "네오폴리탄 마스티프",
-  "뉴펀들랜드",
-  "닥스훈트",
-  "달마시안",
-  "댄디 딘몬트 테리어",
-  "도고 까나리오",
-  "도그 도 보르도",
-  "도베르만 핀셔",
-  "로트와일러",
-  "로디지아 리지백",
-  "로첸",
-  "라사 압소",
-  "라포니안 허더",
-  "레브라도 리트리버",
-  "레이크랜드 테리어",
-  "말티즈",
-  "맨체스터 테리어",
-  "마스티프",
-  "미니어처 불테리어",
-  "미니어처 핀셔",
-  "바센지",
-  "바셋 하운드",
-  "버니즈 마운틴 독",
-  "베를링턴 테리어",
-  "벨기에 말리노이즈",
-  "벨기에 시프도그",
-  "벨기에 테뷰런",
-  "벨지안 그리펀",
-  "벨지안 세퍼드",
-  "보더 콜리",
-  "보더 테리어",
-  "보르조이",
-  "보스롱",
-  "보스턴 테리어",
-  "복서",
-  "부비에 데 플란더스",
-  "불독",
-  "불마스티프",
-  "불테리어",
-  "브뤼셀 그리펀",
-  "브리타니",
-  "브리아드",
-  "블랙 러시안 테리어",
-  "블랙 앤드 탄 쿤하운드",
-  "블러드 하운드",
-  "비글",
-  "비숑 프리제",
-  "비어디드 콜리",
-  "비즐라",
-  "사모예드",
-  "살루키",
-  "서식스 스패니얼",
-  "세인트 버나드",
-  "셔틀랜드 쉽독",
-  "슈나우저",
-  "스무드 폭스 테리어",
-  "스위디시 발훈트",
-  "스카이 테리어",
-  "스코티시 디어하운드",
-  "스코티시 테리어",
-  "스피노 이탈리아노",
-  "스피츠",
-  "스피퍼키",
-  "스페니쉬 그레이하운드",
-  "스페니쉬 마스티프",
-  "스태퍼드셔 불테리어",
-  "시바 이누",
-  "시베리안 허스키",
-  "시추",
-  "실리업 테리어",
-  "실키 테리어",
-  "아나톨리아 세퍼드 독",
-  "아메리칸 스태퍼드셔 테리어",
-  "아메리칸 아키다",
-  "아메리칸 에스키모 독",
-  "아메리칸 워터 스패니얼",
-  "아메리칸 코커 스패니얼",
-  "아메리칸 폭스하운드",
-  "아이리시 레드 앤드 화이트 세터",
-  "아이리시 세터",
-  "아이리시 소프트코티드 휘튼 테리어",
-  "아이리시 울프하운드",
-  "아이리시 워터 스패니얼",
-  "아이리시 테리어",
-  "아키타",
-  "아펜핀셔",
-  "아프간 하운드",
-  "알래스칸 말라뮤트",
-  "에스트렐라 마운틴 독",
-  "에어데일 테리어",
-  "오스트레일리언 실키 테리어",
-  "오스트레일리언 세퍼드",
-  "오스트레일리언 캐틀독",
-  "오스트레일리언 테리어",
-  "오터 하운드",
-  "올드 잉글리시 쉽도그",
-  "와이어 폭스 테리어",
-  "와이어 헤어드 포인팅 그리펀",
-  "와이마라너",
-  "요크셔 테리어",
-  "웰시 코기",
-  "웰시 스프링어 스패니얼",
-  "웰시 테리어",
-  "이비전 하운드",
-  "이탈리안 그레이하운드",
-  "잉글리시 세터",
-  "잉글리시 스프링어 스패니얼",
-  "잉글리시 코커 스패니얼",
-  "잉글리시 토이 스패니얼",
-  "잉글리시 폭스하운드",
-  "저먼 세퍼드 독",
-  "저먼 쇼트헤어드 포인터",
-  "저먼 와이어헤어드 포인터",
-  "저먼 핀셔",
-  "저먼 헌팅 테리어",
-  "진돗개",
-  "차우차우",
-  "차이니스 샤페이",
-  "차이니스 크레스티드",
-  "체서피크 베이 레트리버",
-  "치와와",
-  "카발리에 킹 찰스 스패니얼",
-  "컬리 코티드 레트리버",
-  "케이스 혼트",
-  "케이넌 독",
-  "케언 테리어",
-  "케리 블루 테리어",
-  "코커 스패니얼",
-  "코몬도르",
-  "코튼 드 툴리어",
-  "콜리",
-  "쿠바스",
-  "클럼버 스패니얼",
-  "토이 푸들",
-  "토이 폭스 테리어",
-  "티베탄 마스티프",
-  "티베탄 스패니얼",
-  "티베탄 테리어",
-  "파라오 하운드",
-  "파슨 러셀 테리어",
-  "파피용",
-  "페키니즈",
-  "포메라니안",
-  "포르투갈 워터독",
-  "푸들",
-  "풍산개",
-];
